Remove unused help shell. Slight cleanup and improvement.
