+++ crates/nu-engine/src/filesystem/filesystem_shell.rs
@@ -58,6 +58,10 @@ impl Clone for FilesystemShell {
 }
 
 impl FilesystemShell {
+    fn is_cli(&self) -> bool {
+        matches!(&self.mode, FilesystemShellMode::Cli)
+    }
+
     pub fn basic(mode: FilesystemShellMode) -> Result<FilesystemShell, Error> {
         let path = match std::env::current_dir() {
             Ok(path) => path,
@@ -295,7 +299,7 @@ impl Shell for FilesystemShell {
 
         //Loading local configs in script mode, makes scripts behave different on different
         //filesystems and might therefore surprise users. That's why we only load them in cli mode.
-        if self.mode == FilesystemShellMode::Cli {
+        if self.is_cli() {
             match dunce::canonicalize(self.path()) {
                 Err(e) => {
                     trace!(
