+++ crates/nu-engine/src/shell/mod.rs
@@ -11,7 +11,6 @@ use std::path::{Path, PathBuf};
 use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 
-pub(crate) mod help_shell;
 pub(crate) mod painter;
 pub(crate) mod palette;
 pub(crate) mod shell_args;
